Refactor server sources to add /src folder (to be like client side and to have a better sources organization)
