+++ server/app.js
@@ -4,10 +4,11 @@ const favicon = require('serve-favicon');
 const logger = require('morgan');
 const cookieParser = require('cookie-parser');
 const bodyParser = require('body-parser');
-
-const index = require('./routes/index');
-const jobs = require('./routes/jobs');
 const cors = require('cors');
+
+const index = require('./src/routes/index');
+const jobs = require('./src/routes/jobs');
+const auth = require('./src/routes/auth');
 
 const app = express();
 
@@ -22,10 +23,14 @@ app.use(cookieParser());
 app.use(cors());
 
 // static resources
-app.use(express.static(path.join(__dirname, '..', 'client', 'dist')));
+// FIXME manage better environment variables
+if ('test' !== process.env.NODE_ENV) {
+  app.use(express.static(path.join(__dirname, '..', 'client', 'dist')));
+}
 
 app.use('/', index);
 app.use('/api/jobs', jobs);
+app.use('/auth', auth);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
