+++ server/package.json
@@ -6,6 +6,7 @@
     "start": "node ./bin/www",
     "clean": "rm -rf node_modules",
     "test": "NODE_ENV=test ./node_modules/mocha/bin/mocha --recursive",
+    "test:watch": "NODE_ENV=test ./node_modules/mocha/bin/mocha --recursive --watch",
     "lint": "./node_modules/.bin/eslint --fix --ext .js routes test"
   },
   "dependencies": {
