+++ server/test/auth.spec.js
@@ -1,0 +1,11 @@
+const { describe, it, request, expect } = require('./test-helper')
+
+describe('', function () {
+  it('should...', function () {
+    // given
+
+    // when
+
+    // then
+  })
+})
