+++ server/test/index.spec.js
@@ -1,9 +1,9 @@
 const { describe, it, request, expect } = require('./test-helper')
-const index = require('../app')
+const app = require('../app')
 
 describe('API', function () {
-  it('should have api informations on root', () => {
-    request(index)
+  it('should have api informations on root', (done) => {
+    request(app)
       .get('/')
       .expect('Content-Type', /json/)
       .end(function (err, res) {
@@ -11,6 +11,7 @@ describe('API', function () {
           done(err)
         }
         expect(res.body).to.have.property('azerty')
+        done()
       })
   })
 })
